Server is almost finished (comments todo)

+++ back/src/Application/Actions/User/DeleteUserAction.php
@@ -2,6 +2,7 @@
 
 namespace App\Application\Actions\User;
 
+use App\Domain\User;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -12,8 +13,16 @@ class DeleteUserAction extends UserAction
      */
     public function action(Request $request): Response
     {
+        /** @var User $currentUser */
+        $currentUser = $this->getUser();
+
         $username = $request->get('username');
-        //TODO: verify if it is a user himself who delete it
+
+        if ($username !== $currentUser->getUsername()) {
+            //TODO: 401 instead of 403?
+            throw $this->createAccessDeniedException("You're not authorized to perform this action");
+        }
+
         $user = $this->userRepository->deleteUser($username);
 
         $this->logger->info("Todo of id `{$user->getId()}` was deleted.");
